[Superbeing] update superbeings packages
