+++ src/main/java/com/shield/projectJavaCesi/repository/superbeing/ISuperbeingRepository.java
@@ -1,4 +1,4 @@
-package com.shield.projectJavaCesi.repository;
+package com.shield.projectJavaCesi.repository.superbeing;
 
 import com.shield.projectJavaCesi.entity.superbeing.Superbeing;
 import org.springframework.data.jpa.repository.JpaRepository;
